+++ components/ResultsPage/YourAnswers.tsx
@@ -55,7 +55,7 @@ export const YourAnswers: React.VFC<{
                   />
                 </div>
                 <div className="justify-self-end self-end">
-                  <Link href={`eligibility#${input.key}`}>
+                  <Link href={`questions#${input.key}`}>
                     <a
                       className="ds-underline ds-text-multi-blue-blue70b ds-font-body ds-text-browserh5 ds-leading-33px hover:ds-text-multi-blue-blue50b"
                       aria-label={tsln.resultsEditAriaLabels[input.key]}
