+++ pages/index.tsx
@@ -114,7 +114,7 @@ const Home: NextPage<{ adobeAnalyticsUrl: string }> = ({
             <Button
               text={tsln.startBenefitsEstimator}
               styling="supertask"
-              onClick={(e) => router.push('/eligibility')}
+              onClick={(e) => router.push('/questions')}
               className=" w-auto justify-center mb-4"
               attributes={{
                 [AA_CUSTOMCLICK]: `${AA_BUTTON_CLICK_ATTRIBUTE}:${tsln.startBenefitsEstimator}`,
